WinForm-BinaryDatabase
The program in C# - in the Polish language version. The binary database
that stores personal information. The content of the database is stored
in a binary file. The database is specifically limited to handle up to
10 people. Adding eleven people overwrites the data for the person
entered at the earliest. The application allows you to add / delete /
view and modify data. The application allows you to import and export
data to a text file.